I attempted to fix persistent build errors CS0101 and CS0246.
This commit includes changes I made to resolve two critical build errors:
- CS0101: "The namespace 'yt_dlp_gui' already contains a definition for 'App'"
- CS0246: "The type or namespace name 'ToastNotificationActivatedEventArgsCompat' could not be found"

Here's what I did:
1.  I added `using Microsoft.Toolkit.Uwp.Notifications;` to `yt-dlp-gui/Views/Main.xaml.cs` and ensured related code was uncommented.
2.  I merged methods and properties from `yt-dlp-gui/App/App.Path.cs` into `yt-dlp-gui/App.xaml.cs`.
3.  I removed the file `yt-dlp-gui/App/App.Path.cs`.
4.  I updated the content of `yt-dlp-gui/yt-dlp-gui.csproj` with the version you provided multiple times to ensure its integrity.
5.  I consolidated `yt-dlp-gui/App.xaml` into `yt-dlp-gui/App.xaml.cs` by:
    - Removing the file `App.xaml`.
    - Modifying `App.xaml.cs` to handle application startup in `OnStartup`.
    - Ensuring the `App` class in `App.xaml.cs` is not marked as `partial`.
6.  I attempted to fix CS0246 by explicitly removing and re-adding the `Microsoft.Toolkit.Uwp.Notifications` package.
7.  I performed clean builds by removing `obj` and `bin` folders before each build attempt.

Despite these efforts, both CS0101 and CS0246 errors persist.
It appeared that critical `.csproj` settings like `<EnableWindowsTargeting>true</EnableWindowsTargeting>` and `<RootNamespace>yt_dlp_gui</RootNamespace>` needed to be re-applied before build steps, suggesting an issue with how the build environment handles or maintains the project file's state.
These errors are likely related to the complexities of building WPF projects (which are Windows-specific) in the provided Linux-based environment, especially when `EnableWindowsTargeting` is required. The build tools may not be interpreting the project structure or dependencies correctly under these conditions.

The CS8625 warnings regarding nullable reference types also remain.
Further investigation into the build environment configuration and tooling specific to WPF cross-compilation appears necessary.
