--- yt-dlp-gui/App.xaml
@@ -1,13 +1,0 @@
-﻿<Application x:Class="yt_dlp_gui.App"
-             xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-             xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-             xmlns:local="clr-namespace:yt_dlp_gui"
-             Startup="Application_Startup">
-    <Application.Resources>
-        <ResourceDictionary>
-            <ResourceDictionary.MergedDictionaries>
-                <!-- Theme dictionaries will be added here by ThemeManager -->
-            </ResourceDictionary.MergedDictionaries>
-        </ResourceDictionary>
-    </Application.Resources>
-</Application>
