+++ yt-dlp-gui/yt-dlp-gui.csproj
@@ -3,143 +3,22 @@
   <PropertyGroup>
     <OutputType>WinExe</OutputType>
     <TargetFramework>net6.0-windows</TargetFramework>
-    <RootNamespace>yt_dlp_gui</RootNamespace>
     <UseWPF>true</UseWPF>
     <ApplicationIcon>Resources\logo.ico</ApplicationIcon>
     <Nullable>enable</Nullable>
     <ImplicitUsings>enable</ImplicitUsings>
   </PropertyGroup>
 
-  <PropertyGroup>
-    <EnableDefaultApplicationDefinition>false</EnableDefaultApplicationDefinition>
-    <EnableDefaultPageItems>false</EnableDefaultPageItems>
-    <EnableDefaultCompileItems>false</EnableDefaultCompileItems>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <ApplicationDefinition Include="App.xaml">
-      <Generator>MSBuild:Compile</Generator>
-      <XamlRuntime>Wpf</XamlRuntime>
-    </ApplicationDefinition>
-  </ItemGroup>
-
-  <ItemGroup>
-    <Compile Include="App.xaml.cs">
-      <DependentUpon>App.xaml</DependentUpon>
-      <SubType>Code</SubType>
-    </Compile>
-    <Compile Include="App/App.Path.cs">
-      <DependentUpon>App.xaml</DependentUpon>
-    </Compile>
-    <Compile Include="App/YtdlpUpdater.cs">
-      <DependentUpon>App.xaml</DependentUpon>
-    </Compile>
-    <Compile Include="AssemblyInfo.cs" />
-    <Compile Include="Controls/ActualSize.cs" />
-    <Compile Include="Controls/BrowserBehavior.cs" />
-    <Compile Include="Controls/Duration.cs" />
-    <Compile Include="Controls/Filesize.cs" />
-    <Compile Include="Controls/Icons.xaml.cs">
-      <DependentUpon>Icons.xaml</DependentUpon>
-    </Compile>
-    <Compile Include="Controls/Menu.xaml.cs">
-      <DependentUpon>Menu.xaml</DependentUpon>
-    </Compile>
-    <Compile Include="Controls/TextBoxNumber.cs" />
-    <Compile Include="Controls/TextEditor.xaml.cs">
-      <DependentUpon>TextEditor.xaml</DependentUpon>
-    </Compile>
-    <Compile Include="Controls/TextView.xaml.cs">
-      <DependentUpon>TextView.xaml</DependentUpon>
-    </Compile>
-    <Compile Include="Converters/NullOrEmptyToVisibilityConverter.cs" />
-    <Compile Include="Libs/Util.Env.cs" />
-    <Compile Include="Libs/Util.PropertyCopy.cs" />
-    <Compile Include="Libs/Util.Resources.cs" />
-    <Compile Include="Libs/Util.Shell.cs" />
-    <Compile Include="Libs/Util.Tools.cs" />
-    <Compile Include="Libs/Util.UrlValid.cs" />
-    <Compile Include="Libs/Web.cs" />
-    <Compile Include="Libs/Yaml.cs" />
-    <Compile Include="Libs/YamlDescription.cs" />
-    <Compile Include="Models/Chapters.cs" />
-    <Compile Include="Models/Config.cs" />
-    <Compile Include="Models/Cookies.cs" />
-    <Compile Include="Models/DownloadItem.cs" />
-    <Compile Include="Models/Format.cs" />
-    <Compile Include="Models/Lang.cs" />
-    <Compile Include="Models/Subs.cs" />
-    <Compile Include="Models/Thumb.cs" />
-    <Compile Include="Models/Video.cs" />
-    <Compile Include="Models/VTask.cs" />
-    <Compile Include="Themes/CustomUI.xaml.cs">
-      <DependentUpon>CustomUI.xaml</DependentUpon>
-    </Compile>
-    <Compile Include="Themes/DarkTheme.xaml.cs">
-      <DependentUpon>DarkTheme.xaml</DependentUpon>
-    </Compile>
-    <Compile Include="ViewModels/Main.cs" />
-    <Compile Include="Views/About.xaml.cs">
-      <DependentUpon>About.xaml</DependentUpon>
-    </Compile>
-    <Compile Include="Views/Main.xaml.cs">
-      <DependentUpon>Main.xaml</DependentUpon>
-    </Compile>
-    <Compile Include="Views/Release.xaml.cs">
-      <DependentUpon>Release.xaml</DependentUpon>
-    </Compile>
-    <Compile Include="Wrappers/DLP.cs" />
-    <Compile Include="Wrappers/FFMPEG.cs" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <Page Include="Controls/Icons.xaml">
-      <SubType>Designer</SubType>
-      <Generator>MSBuild:Compile</Generator>
-    </Page>
-    <Page Include="Controls/Menu.xaml">
-      <SubType>Designer</SubType>
-      <Generator>MSBuild:Compile</Generator>
-    </Page>
-    <Page Include="Controls/TextEditor.xaml">
-      <SubType>Designer</SubType>
-      <Generator>MSBuild:Compile</Generator>
-    </Page>
-    <Page Include="Controls/TextView.xaml">
-      <SubType>Designer</SubType>
-      <Generator>MSBuild:Compile</Generator>
-    </Page>
-    <Page Include="Themes/CustomUI.xaml">
-      <SubType>Designer</SubType>
-      <Generator>MSBuild:Compile</Generator>
-    </Page>
-    <Page Include="Themes/DarkTheme.xaml">
-      <SubType>Designer</SubType>
-      <Generator>MSBuild:Compile</Generator>
-    </Page>
-    <Page Include="Views/About.xaml">
-      <SubType>Designer</SubType>
-      <Generator>MSBuild:Compile</Generator>
-    </Page>
-    <Page Include="Views/Main.xaml">
-      <SubType>Designer</SubType>
-      <Generator>MSBuild:Compile</Generator>
-    </Page>
-    <Page Include="Views/Release.xaml">
-      <SubType>Designer</SubType>
-      <Generator>MSBuild:Compile</Generator>
-    </Page>
-  </ItemGroup>
-
   <ItemGroup>
     <None Remove="Resources\logo.ico" />
     <None Remove="Resources\xshd\desc.xshd" />
     <None Remove="Resources\xshd\url.xshd" />
+  </ItemGroup>
+
+  <ItemGroup>
     <Resource Include="Resources\logo.ico" />
-    <Resource Include="Resources\xshd\desc.xshd" />
-    <Resource Include="Resources\xshd\url.xshd">
-      <CustomToolNamespace></CustomToolNamespace>
-    </Resource>
+    <EmbeddedResource Include="Resources\xshd\desc.xshd" />
+    <EmbeddedResource Include="Resources\xshd\url.xshd" />
   </ItemGroup>
 
   <ItemGroup>
@@ -147,17 +26,39 @@
     <PackageReference Include="Markdig" Version="0.41.2" />
     <PackageReference Include="MdXaml_migfree" Version="1.15.0" />
     <PackageReference Include="Microsoft.Toolkit.Uwp.Notifications" Version="7.1.3" />
-    <PackageReference Include="Microsoft.Xaml.Behaviors.Wpf" Version="1.1.135" />
+    <PackageReference Include="Microsoft.Xaml.Behaviors.Wpf" Version="1.1.77" />
     <PackageReference Include="Newtonsoft.Json" Version="13.0.3" />
     <PackageReference Include="PropertyChanged.Fody" Version="4.1.0" />
     <PackageReference Include="SharpClipboard" Version="3.5.2" />
     <PackageReference Include="Swordfish.NET.CollectionsV3" Version="3.3.15" />
-    <PackageReference Include="YamlDotNet" Version="16.3.0" />
+    <PackageReference Include="System.Resources.ResourceManager" Version="4.3.0" />
+    <PackageReference Include="YamlDotNet" Version="16.0.0" />
   </ItemGroup>
 
   <ItemGroup>
-    <None Update="FodyWeavers.xml">
-      <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
+    <Compile Update="Properties\Resources.Designer.cs">
+      <DesignTime>True</DesignTime>
+      <AutoGen>True</AutoGen>
+      <DependentUpon>Resources.resx</DependentUpon>
+    </Compile>
+    <Compile Update="Properties\Settings.Designer.cs">
+      <DesignTimeSharedInput>True</DesignTimeSharedInput>
+      <AutoGen>True</AutoGen>
+      <DependentUpon>Settings.settings</DependentUpon>
+    </Compile>
+  </ItemGroup>
+
+  <ItemGroup>
+    <EmbeddedResource Update="Properties\Resources.resx">
+      <Generator>ResXFileCodeGenerator</Generator>
+      <LastGenOutput>Resources.Designer.cs</LastGenOutput>
+    </EmbeddedResource>
+  </ItemGroup>
+
+  <ItemGroup>
+    <None Update="Properties\Settings.settings">
+      <Generator>SettingsSingleFileGenerator</Generator>
+      <LastGenOutput>Settings.Designer.cs</LastGenOutput>
     </None>
   </ItemGroup>
